shipping-service/pyproject.toml
[tool.poetry]
name = "shipping-service"
version = "0.1.0"
description = ""
authors = ["Marten Elmgren <elmgren@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.1"
alembic = "^1.7.3"
SQLAlchemy = "^1.4.25"
uvicorn = "^0.15.0"
psycopg2 = "^2.9.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
python-dotenv = "^0.19.0"
pylint = "^2.11.1"
flake8 = "^3.9.2"
psycopg2-binary = "^2.9.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
